Octree Display
Added function to output octree structure to console. Had it originally in imgui, but it was making the program run much slower. since the octree kept getting recreated. I also changed the octree to be cleared rather than deleted. may change this again but for now it means the output works. because otherwise it was getting deleted and recreated very frame so sometimes the output would be null